Fix error in validateUserSession
Error was:
Argument `where` of type security_usersWhereUniqueInput needs at least one of `UserID` or `UserID_SiteID` arguments.

Fix was:
Changed the query from findUnique to findFirst since we're not using a unique identifier in the where clause. This should resolve the error you're seeing. The findFirst method will return the first matching record based on the provided conditions, which is what we want in this case.